requirements.txt
# Core AutoGen and Multi-Agent Dependencies
pyautogen>=0.2.0
openai>=1.0.0
anthropic>=0.7.0

# Web Browsing and Automation
playwright>=1.40.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Code Execution and Sandbox
docker>=6.1.0
subprocess32>=3.5.4

# Data Processing and Analysis
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework for Interface
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
gradio>=4.0.0

# Configuration Management
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and Storage
sqlalchemy>=2.0.0
redis>=5.0.0
pymongo>=4.6.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
coverage>=7.3.0

# Code Quality and Linting
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0
aiofiles>=23.2.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Development Tools
jupyter>=1.0.0
ipython>=8.17.0
notebook>=7.0.0

# File Processing
python-multipart>=0.0.6
Pillow>=10.1.0

# Version Control Integration
GitPython>=3.1.40

# Environment and System
psutil>=5.9.0
python-dotenv>=1.0.0

# Network and HTTP
httpx>=0.25.0
websockets>=12.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Additional Utilities
rich>=13.7.0
typer>=0.9.0
click>=8.1.0
tqdm>=4.66.0 